Replaced original HL7 examples with ORM^O01 example (HL7 v2.6); Removed unnecessary includes in hl7_server; Changed some data in HL7 message
