+++ examples/cpp/hl7/hl7_client.cpp
@@ -24,21 +24,17 @@
 
 #include "finalmq/remoteentity/RemoteEntityContainer.h"
 #include "finalmq/remoteentity/RemoteEntityFormatHl7.h"
-#include "finalmq/remoteentity/RemoteEntityFormatJson.h"
 #include "finalmq/logger/Logger.h"
 #include "finalmq/variant/VariantValueStruct.h"
-#include "finalmq/variant/VariantValues.h"
-#include "finalmq/protocols/ProtocolMqtt5Client.h"
-#include "finalmq/helpers/Executor.h"
 #include "finalmq/protocols/ProtocolDelimiterX.h"
-#include "finalmq/helpers/ProtothreadRequestReply.h"
 
-// the definition of the messages is in the file hl7dictionary.2.7.1.js.fmq
-#include "hl7dictionary.2.7.1.js.fmq.h"
+// the definition of the messages is in the file hl7dictionary.2.6.js.fmq
+#include "hl7dictionary.2.6.js.fmq.h"
 
 #include <iostream>
 #include <thread>
-#include <algorithm>
+#include <iomanip>
+#include <map>
 
 // the modulename is needed for the logger streams (streamDebug, streamInfo, streamWarning, streamError, streamCritical, streamFatal)
 #define MODULENAME  "hl7_server"
@@ -52,145 +48,259 @@ class EntityClient : public RemoteEntity
 public:
     EntityClient()
     {
-        // register peer events to see when a remote entity connects or disconnects.
-        registerPeerEvent([](PeerId peerId, const SessionInfo& session, EntityId entityId, PeerEvent peerEvent, bool incoming) {
-            streamInfo << "peer event " << peerEvent.toString();
-        });
+        // RECEIVE ack
+        registerCommand<hl7::ACK>([this](const RequestContextPtr& requestContext, const std::shared_ptr<hl7::ACK>& ack) {
+            assert(ack);
+            std::cout << "Received ACK: " << ack->msh.messageControlId << std::endl;
 
-
-
-        registerCommand<hl7::SSU_U03>([this](const RequestContextPtr& requestContext, const std::shared_ptr<hl7::SSU_U03>& request) {
-            assert(request);
-
-//            std::cout << "Result: " << request->equ.alertLevel.alternateIdentifier << std::endl;
-
-            std::cout << ".";
+            {
+                std::lock_guard<std::mutex> lock(m_mutex);
+                auto it = m_requests.find(ack->msh.messageControlId);
+                if (it != m_requests.end())
+                {
+                    m_requests.erase(it);   // remove acknowledged message
+                }
+            }
         });
     }
+
+    // SEND
+    void sendHl7Message(finalmq::StructBasePtr& msg)
+    {
+        hl7::MSH& msh = reinterpret_cast<hl7::ACK*>(msg.get())->msh;    // finalmq does not have a base class for HL7 Messages, yet, so I use the ACK to get MSH
+        
+        {
+            std::lock_guard<std::mutex> lock(m_mutex);
+            m_id++;
+            std::string idAsString = std::to_string(m_id);
+            msh.messageControlId = idAsString;
+            m_requests[idAsString] = msg;   // keep not acknowledged messages
+            std::cout << "Sending MSG: " << msh.messageControlId << std::endl;
+        }
+
+        sendEvent(m_peerId, *msg);
+    }    
+
+    void connected(PeerId peerId)
+    {
+        m_peerId = peerId;
+
+        // collect all unacknowledged messages
+        std::vector<finalmq::StructBasePtr> messagesToResend;
+
+        {
+            std::lock_guard<std::mutex> lock(m_mutex);
+            for (auto element : m_requests)
+            {
+                messagesToResend.push_back(element.second);
+            }
+        }
+
+        // resend all unacknowledged messages
+        for (auto msg : messagesToResend)
+        {
+            sendEvent(m_peerId, *msg);
+        }
+    }
+
+    std::map<std::string, finalmq::StructBasePtr> m_requests{};
+    int m_id{ 0 };
+    PeerId m_peerId{};
+    std::mutex  m_mutex{};
 };
+
+class HL7Client
+{
+public:
+    HL7Client(std::shared_ptr<EntityClient> entityClientPtr)
+    {
+        m_entityClient = entityClientPtr;
+
+        // display log traces
+        Logger::instance().registerConsumer([](const LogContext& context, const char* text) {
+            std::cout << context.filename << "(" << context.line << ") " << text << std::endl;
+            });
+
+        // Initialize entity container. Entities can be added with registerEntity().
+        // Entities are like remote objects, but they can be at the same time client and server.
+        // This means, an entity can send (client) and receive (server) a request command.
+        entityContainer.init();
+
+        // run entity container in separate thread
+        m_thread = std::thread([this]() {
+            entityContainer.run();
+            });
+
+
+        // register lambda for connection events to see when a network node connects or disconnects.
+        entityContainer.registerConnectionEvent([this](const SessionInfo& session, ConnectionEvent connectionEvent) {
+            const ConnectionData connectionData = session.getConnectionData();
+            streamInfo << "connection event at " << connectionData.endpoint
+                << " remote: " << connectionData.endpointPeer
+                << " event: " << connectionEvent.toString();
+
+            switch (connectionEvent)
+            {
+            case ConnectionEvent::CONNECTIONEVENT_DISCONNECTED:
+                doConnect();    // do an automatic reconnect, in case the connection gets lost
+                break;
+            }
+        });
+
+        entityContainer.registerEntity(entityClientPtr, "Hl7Entity");
+
+        doConnect(); // connect to the server
+    }
+
+    void doConnect()
+    {
+        SessionInfo sessionClient = entityContainer.connect("tcp://localhost:7000:delimiter_x:hl7", { {}, {},
+            VariantStruct{ {ProtocolDelimiterX::KEY_DELIMITER, "\x1C\x0D"} },
+            VariantStruct{  {RemoteEntityFormatHl7::PROPERTY_NAMESPACE, "hl7"},
+                            {RemoteEntityFormatHl7::PROPERTY_ENTITY, "Hl7Entity"},
+                            {RemoteEntityFormatHl7::PROPERTY_MESSAGESTART, "\x0B"} } });
+
+
+        PeerId peerId = m_entityClient->createPublishPeer(sessionClient, "MyService");
+        m_entityClient->connected(peerId);
+    }
+
+    ~HL7Client()
+    {
+        // release the thread
+        entityContainer.terminatePollerLoop();
+        m_thread.join();
+    }
+
+private:
+    std::thread m_thread;
+    RemoteEntityContainer entityContainer;
+    std::shared_ptr<EntityClient> m_entityClient{};
+};
+
+
+
+
+static finalmq::StructBasePtr create_ORM_O01()
+{
+    auto msgPtr = std::make_shared< hl7::ORM_O01>();
+
+    // Date Time Now
+    auto now = std::chrono::system_clock::now();
+    auto in_time_t = std::chrono::system_clock::to_time_t(now);
+    std::stringstream ssDateTime;
+    ssDateTime << std::put_time(std::localtime(&in_time_t), "%Y%m%d%H%M%S");
+    std::string dateTimeStr = ssDateTime.str();
+    std::stringstream ssDate;
+    ssDate << std::put_time(std::localtime(&in_time_t), "%Y%m%d");
+    std::string dateStr = ssDate.str();
+
+    // MSH
+    msgPtr->msh.sendingApplication = hl7::HD("MyNamespace", "MyApplicationId", hl7::UniversalIdType());
+    msgPtr->msh.sendingFacility = hl7::HD("MyNamespace", "MyFacilityId", hl7::UniversalIdType());
+    msgPtr->msh.receivingApplication = hl7::HD("YourNamespace", "YourApplicationId", hl7::UniversalIdType());
+    msgPtr->msh.receivingFacility = hl7::HD("YourNamespace", "YourFacilityId", hl7::UniversalIdType());
+    msgPtr->msh.date_timeOfMessage = dateTimeStr;
+    msgPtr->msh.messageControlId = "1234567890";
+    msgPtr->msh.processingId.processingId = hl7::ProcessingId::Enum::Production;
+    msgPtr->msh.versionId.versionId = hl7::VersionId::Enum::Release2_6;
+    msgPtr->msh.characterSet = { "8859/1" };
+
+    // Patient Start
+    auto patient = std::make_shared<hl7::PATIENT_13>();
+
+    // PID
+    patient->pid.patientIdentifierList.resize(1);
+    patient->pid.patientIdentifierList[0].idNumber = "ID 000123/24";
+    patient->pid.patientName.resize(1);
+    patient->pid.patientName[0].familyName.surname = "Testperson";
+    patient->pid.patientName[0].givenName = "Testi";
+    patient->pid.date_timeOfBirth = "19700101";
+    patient->pid.administrativeSex = "W";
+
+    // PV1
+    auto patientVisit = std::make_shared<hl7::PATIENT_VISIT_2>();
+    patientVisit->pv1.patientClass = "A";
+    patientVisit->pv1.assignedPatientLocation.pointOfCare = "DKD";
+    patientVisit->pv1.referringDoctor.resize(1);
+    patientVisit->pv1.referringDoctor[0].idNumber = "DocID";
+    patientVisit->pv1.temporaryLocation.pointOfCare = "POCCT01";
+    patientVisit->pv1.visitNumber.idNumber = "ID 000123/24";
+    patientVisit->pv1.admitDate_time = dateStr;
+    patient->patient_visit = patientVisit;
+
+    msgPtr->patient = patient;
+
+    // Order Start
+    hl7::ORDER_26 order;
+
+    // ORC
+    hl7::ORC orc;
+    orc.orderControl = "NW";
+    orc.placerOrderNumber.entityIdentifier = "ID00012324000111";
+    orc.placerGroupNumber.entityIdentifier = "BCT";
+    orc.responseFlag = "0";
+    //orc.quantity_timing.resize(1);
+    //orc.quantity_timing[0].interval.repeatPattern = hl7::RepeatPattern::Empty;
+    orc.date_timeOfTransaction = dateTimeStr;
+    orc.enteredBy.resize(1);
+    orc.enteredBy[0].idNumber = "DocID";
+    orc.orderingProvider.resize(1);
+    orc.orderingProvider[0].idNumber = "DocID";
+    orc.enteringDevice.identifier = "BCT";
+    order.orc = orc;
+
+
+    // OBR
+    auto obr = std::make_shared<hl7::OBR>();
+    obr->placerOrderNumber.entityIdentifier = "ID00012324000111";
+    obr->placerOrderNumber.namespaceId = "1"; // notwendig?
+    obr->universalServiceIdentifier.identifier = "CT_KNEE_L";
+    obr->universalServiceIdentifier.alternateIdentifier = "CT";
+    obr->placerField1 = "-";
+    //obr->quantity_timing.resize(1);
+    //obr->quantity_timing[0].startDate_time = dateTimeStr;
+    auto obrEtc = hl7::OBR_RQD_RQ1_RXO_ODS_ODT_1();
+    auto orderDetail = std::make_shared<hl7::ORDER_DETAIL_1>();
+    obrEtc.obr = obr;
+    orderDetail->obr_rqd_rq1_rxo_ods_odt = { obrEtc };
+
+    // OBX
+    hl7::OBSERVATION_2 observation1;
+    hl7::OBSERVATION_2 observation2;
+    observation1.obx.setId_Obx = "1";
+    observation1.obx.valueType = "TX";
+    observation1.obx.observationIdentifier.identifier = "WEIGHT";
+    observation1.obx.observationValue = { "70" };
+    observation1.obx.units.identifier = "kg";
+    observation2.obx.setId_Obx = "2";
+    observation2.obx.valueType = "TX";
+    observation2.obx.observationIdentifier.identifier = "HEIGHT";
+    observation2.obx.observationValue = { "165" };
+    observation2.obx.units.identifier = "cm";
+    orderDetail->observation = { observation1, observation2 };
+
+    order.order_detail = orderDetail;
+    msgPtr->order = { order };
+
+    return msgPtr;
+}
 
 
 
 int main()
 {
-    // display log traces
-    Logger::instance().registerConsumer([] (const LogContext& context, const char* text) {
-        std::cout << context.filename << "(" << context.line << ") " << text << std::endl;
-    });
+    std::shared_ptr<EntityClient> entityClientPtr = std::make_shared<EntityClient>();
+    HL7Client client(entityClientPtr);
 
-    // Create and initialize entity container. Entities can be added with registerEntity().
-    // Entities are like remote objects, but they can be at the same time client and server.
-    // This means, an entity can send (client) and receive (server) a request command.
-    RemoteEntityContainer entityContainer;
+    for (int i = 0; i < 10; i++)
+	{
+        finalmq::StructBasePtr msg = create_ORM_O01();
 
-    entityContainer.init();
-    // run entity container in separate thread
-    std::thread thread([&entityContainer] () {
-        entityContainer.run();
-    });
+        // send from outside
+        entityClientPtr->sendHl7Message(msg);
+	}
 
-
-    // register lambda for connection events to see when a network node connects or disconnects.
-    entityContainer.registerConnectionEvent([] (const SessionInfo& session, ConnectionEvent connectionEvent) {
-        const ConnectionData connectionData = session.getConnectionData();
-        streamInfo << "connection event at " << connectionData.endpoint
-                  << " remote: " << connectionData.endpointPeer
-                  << " event: " << connectionEvent.toString();
-    });
-
-    // Create client entity and register it at the entityContainer
-    // note: multiple entities can be registered.
-    IRemoteEntityPtr entityClientPtr = std::make_shared<EntityClient>();
-    IRemoteEntity& entityClient = *entityClientPtr;
-
-    entityContainer.registerEntity(entityClientPtr, "Hl7Entity");
-
-    // register peer events to see when a remote entity connects or disconnects.
-    entityClient.registerPeerEvent([] (PeerId peerId, const SessionInfo& session, EntityId entityId, PeerEvent peerEvent, bool incoming) {
-        streamInfo << "peer event " << peerEvent.toString();
-    });
-
-    // connect to port 7777 with simple framing protocol ProtocolHeaderBinarySize (4 byte header with the size of payload).
-    // content type in payload: protobuf
-    // note: Also multiple connects are possible.
-    // And by the way, also bind()s are possible. An EntityContainer can be client and server at the same time.
-    // A client can be started before the server is started. The connect is been retried in the background till the server
-    // becomes available. Use the ConnectProperties to change the reconnect properties
-    // (default is: try to connect every 5s forever till the server becomes available).
-    //SessionInfo sessionClient = entityContainer.connect("tcp://localhost:7777:headersize:protobuf");
-    //SessionInfo sessionClient = entityContainer.connect("ipc://my_uds:headersize:protobuf");
-
-    SessionInfo sessionClient = entityContainer.connect("tcp://localhost:7000:delimiter_x:hl7", { {}, {}, 
-        VariantStruct{ {ProtocolDelimiterX::KEY_DELIMITER, "\x1C\x0D"} },
-        VariantStruct{  {RemoteEntityFormatHl7::PROPERTY_NAMESPACE, "hl7"},
-                        {RemoteEntityFormatHl7::PROPERTY_ENTITY, "Hl7Entity"},
-                        {RemoteEntityFormatHl7::PROPERTY_MESSAGESTART, "\x0B"} } });
-
-
-    //SessionInfo sessionClient = entityContainer.connect("tcp://localhost:7001:delimiter_lf:json", { {}, {}, {},
-    //    VariantStruct{  {RemoteEntityFormatJson::PROPERTY_SERIALIZE_ENUM_AS_STRING, true},
-    //                    {RemoteEntityFormatJson::PROPERTY_SERIALIZE_SKIP_DEFAULT_VALUES, true} } });
-
-
-    //// if you want to use mqtt5 -> connect to broker
-    //SessionInfo sessionClient = entityContainer.connect("tcp://localhost:1883:mqtt5client:json", { {},{},
-    //    VariantStruct{  //{ProtocolMqtt5Client::KEY_USERNAME, std::string("")},
-    //                    //{ProtocolMqtt5Client::KEY_PASSWORD, std::string("")},
-    //                    {ProtocolMqtt5Client::KEY_SESSIONEXPIRYINTERVAL, 300},
-    //                    {ProtocolMqtt5Client::KEY_KEEPALIVE, 20},
-    //    } });
-
-    // connect entityClient to remote server entity "MyService" with the created TCP session.
-    // The returned peerId identifies the peer entity.
-    // The peerId will be used for sending commands to the peer (requestReply(), sendEvent())
-    //PeerId peerId = entityClient.connect(sessionClient, "MyService", [](PeerId peerId, Status status) {
-    //    streamInfo << "connect reply: " << status.toString();
-    //    });
-
-    PeerId peerId = entityClient.createPublishPeer(sessionClient, "MyService");
-
-
-    hl7::SSU_U03 msg;
-    msg.msh.countryCode = "de";
-    msg.uac = std::make_shared<hl7::UAC>();
-    msg.uac->userAuthenticationCredential.typeOfData = hl7::MimeTypes::AudioData;
-    msg.sft.resize(3);
-    msg.sft[0].softwareBinaryId = "world";
-    msg.specimen_container.resize(4);
-    msg.specimen_container[0].sac.positionInTray.value1 = "hey";
-    msg.specimen_container[0].sac.specimenSource = "hh";
-    msg.specimen_container[0].sac.carrierIdentifier.entityIdentifier = "uu";
-    msg.specimen_container[0].sac.carrierIdentifier.universalId = "bbb";
-    msg.specimen_container[0].obx.resize(2);
-    msg.specimen_container[0].specimen.resize(3);
-    msg.specimen_container[0].specimen[0].spm.accessionId.resize(1);
-    msg.specimen_container[0].specimen[0].spm.accessionId[0].idNumber = "ggg";
-    msg.specimen_container[0].specimen[0].spm.containerCondition.alternateText = "tt";
-    msg.specimen_container[0].specimen[0].spm.containerCondition.nameOfAlternateCodingSystem = hl7::CodingSystem::AstmE1238_E1467Universal;
-    msg.specimen_container[0].specimen[0].obx.resize(5);
-
-    for (int i = 0; i < 10000; i++)
-    {
-//        entityClient.sendEvent(peerId, msg);
-
-        entityClient.requestReply<hl7::SSU_U03>(peerId,
-            msg,
-            [](PeerId peerId, Status status, const std::shared_ptr<hl7::SSU_U03>& reply) {
-                if (reply)
-                {
-                    std::cout << ".";
-                }
-                else
-                {
-                    std::cout << "REPLY error: " << status.toString() << std::endl;
-                }
-            });
-    }
-
-    std::this_thread::sleep_for(std::chrono::milliseconds(10000000));
-
-    // release the thread
-    entityContainer.terminatePollerLoop();
-    thread.join();
+    std::this_thread::sleep_for(std::chrono::milliseconds(1000000));
     return 0;
 }
